Rename scanner.testing to scanner.servlets because that's what it should be.
