+++ src/main/java/com/google/cloud/security/scanner/servlets/DesiredStateEnforcerApp.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.security.scanner.testing;
+package com.google.cloud.security.scanner.servlets;
 
 import com.google.appengine.api.utils.SystemProperty;
 import com.google.cloud.dataflow.sdk.io.TextIO;
@@ -22,40 +22,56 @@ import com.google.cloud.dataflow.sdk.options.DataflowPipelineOptions;
 import com.google.cloud.dataflow.sdk.options.PipelineOptions;
 import com.google.cloud.dataflow.sdk.options.PipelineOptionsFactory;
 import com.google.cloud.dataflow.sdk.runners.BlockingDataflowPipelineRunner;
-import com.google.cloud.security.scanner.pipelines.ExportedServiceAccountKeyRemover;
+import com.google.cloud.security.scanner.pipelines.DesiredStateEnforcer;
+import com.google.cloud.security.scanner.sources.GCSFilesSource;
 import com.google.common.base.Preconditions;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.security.GeneralSecurityException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /** Handler class for the Dataflow local runner test endpoint. */
-public class UserManagedKeysApp extends HttpServlet {
+public class DesiredStateEnforcerApp extends HttpServlet {
 
+  /**
+   * Handler for the GET request to this app.
+   * @param req The request object.
+   * @param resp The response object.
+   * @throws IOException Thrown if there's an error reading from one of the APIs.
+   */
   @Override
   public void doGet(HttpServletRequest req, HttpServletResponse resp)
       throws IOException {
     PrintWriter out = resp.getWriter();
+    String org = System.getenv("POLICY_SCANNER_ORG_NAME");
     String orgId = System.getenv("POLICY_SCANNER_ORG_ID");
+    String inputRepositoryUrl = System.getenv("POLICY_SCANNER_INPUT_REPOSITORY_URL");
     String sinkUrl = System.getenv("POLICY_SCANNER_SINK_URL");
     String dataflowTmpBucket = System.getenv("POLICY_SCANNER_DATAFLOW_TMP_BUCKET");
     String stagingLocation = "gs://" + dataflowTmpBucket + "/dataflow_tmp";
     boolean executeOnCloud = Boolean.valueOf(System.getenv("POLICY_SCANNER_EXECUTE_ON_CLOUD"));
 
+    Preconditions.checkNotNull(org);
     Preconditions.checkNotNull(orgId);
+    Preconditions.checkNotNull(inputRepositoryUrl);
     Preconditions.checkNotNull(sinkUrl);
     Preconditions.checkNotNull(dataflowTmpBucket);
-
+    GCSFilesSource source;
+    try {
+      source = new GCSFilesSource(inputRepositoryUrl, org);
+    } catch (GeneralSecurityException e) {
+      throw new IOException("SecurityException: Cannot create GCSFileSource");
+    }
     PipelineOptions options;
     if (executeOnCloud) {
       options = getCloudExecutionOptions(stagingLocation);
     } else {
       options = getLocalExecutionOptions();
     }
-
-    new ExportedServiceAccountKeyRemover(options, orgId)
-        .attachSink(TextIO.Write.named("Write output messages").to(sinkUrl))
+    new DesiredStateEnforcer(options, source, orgId)
+        .attachSink(TextIO.Write.named("Write messages to GCS").to(sinkUrl))
         .run();
     out.println("Test passed! The output was written to GCS");
   }
