+++ src/main/java/com/google/cloud/security/scanner/servlets/LiveStateCheckerRunner.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.security.scanner.testing;
+package com.google.cloud.security.scanner.servlets;
 
 import com.google.appengine.api.utils.SystemProperty;
 import com.google.cloud.dataflow.sdk.io.TextIO;
@@ -22,33 +22,24 @@ import com.google.cloud.dataflow.sdk.options.DataflowPipelineOptions;
 import com.google.cloud.dataflow.sdk.options.PipelineOptions;
 import com.google.cloud.dataflow.sdk.options.PipelineOptionsFactory;
 import com.google.cloud.dataflow.sdk.runners.BlockingDataflowPipelineRunner;
-import com.google.cloud.security.scanner.pipelines.LiveStateChecker;
+import com.google.cloud.security.scanner.pipelines.OnDemandLiveStateChecker;
 import com.google.cloud.security.scanner.sources.GCSFilesSource;
 import com.google.common.base.Preconditions;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.security.GeneralSecurityException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
-/** Handler class for the Dataflow local runner test endpoint. */
-public class LiveStateCheckerApp extends HttpServlet {
+/**
+ * For running the Dataflow pipeline directly using the Dataflow SDK.
+ */
+public class LiveStateCheckerRunner {
 
   /**
-   * Handler for the GET request to this app.
-   * @param req The request object.
-   * @param resp The response object.
+   * Main function for the runner.
+   * @param args The args this program was called with.
    * @throws IOException Thrown if there's an error reading from one of the APIs.
    */
-  @Override
-  public void doGet(HttpServletRequest req, HttpServletResponse resp)
-      throws IOException {
-    PrintWriter out = resp.getWriter();
+  public static void main(String[] args) throws IOException, GeneralSecurityException{
     String org = System.getenv("POLICY_SCANNER_ORG_NAME");
-    String orgId = System.getenv("POLICY_SCANNER_ORG_ID");
     String inputRepositoryUrl = System.getenv("POLICY_SCANNER_INPUT_REPOSITORY_URL");
     String sinkUrl = System.getenv("POLICY_SCANNER_SINK_URL");
     String dataflowTmpBucket = System.getenv("POLICY_SCANNER_DATAFLOW_TMP_BUCKET");
@@ -56,11 +47,9 @@ public class LiveStateCheckerApp extends HttpServlet {
     boolean executeOnCloud = Boolean.valueOf(System.getenv("POLICY_SCANNER_EXECUTE_ON_CLOUD"));
 
     Preconditions.checkNotNull(org);
-    Preconditions.checkNotNull(orgId);
     Preconditions.checkNotNull(inputRepositoryUrl);
     Preconditions.checkNotNull(sinkUrl);
     Preconditions.checkNotNull(dataflowTmpBucket);
-
     GCSFilesSource source;
     try {
       source = new GCSFilesSource(inputRepositoryUrl, org);
@@ -73,21 +62,12 @@ public class LiveStateCheckerApp extends HttpServlet {
     } else {
       options = getLocalExecutionOptions();
     }
-
-    String datetimestamp = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss").format(new Date());
-    new LiveStateChecker(options, source, orgId)
-        .attachSink(TextIO.Write.named("Write messages to GCS").to(sinkUrl + "-" + datetimestamp))
+    new OnDemandLiveStateChecker(options, source)
+        .attachSink(TextIO.Write.named("Write messages to GCS").to(sinkUrl))
         .run();
-
-    String outputBucket = sinkUrl.replaceAll("gs://", "");
-    outputBucket = outputBucket.substring(0, outputBucket.lastIndexOf('/'));
-    String outputBucketLink = "https://console.cloud.google.com/storage/browser/" + outputBucket;
-    String outputPage = "<b>Test passed!</b><br><br>"
-                        + "The output was written to GCS: <a href='%s'>" + "output file" + "</a>";
-    out.println(String.format(outputPage, outputBucketLink));
   }
 
-  private PipelineOptions getCloudExecutionOptions(String stagingLocation) {
+  private static PipelineOptions getCloudExecutionOptions(String stagingLocation) {
     DataflowPipelineOptions options = PipelineOptionsFactory.as(DataflowPipelineOptions.class);
     options.setProject(SystemProperty.applicationId.get());
     options.setStagingLocation(stagingLocation);
@@ -95,7 +75,7 @@ public class LiveStateCheckerApp extends HttpServlet {
     return options;
   }
 
-  private PipelineOptions getLocalExecutionOptions() {
+  private static PipelineOptions getLocalExecutionOptions() {
     return PipelineOptionsFactory.create();
   }
 }
