+++ src/main/java/com/google/cloud/security/scanner/servlets/UserManagedKeysApp.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.security.scanner.testing;
+package com.google.cloud.security.scanner.servlets;
 
 import com.google.appengine.api.utils.SystemProperty;
 import com.google.cloud.dataflow.sdk.io.TextIO;
@@ -22,52 +22,45 @@ import com.google.cloud.dataflow.sdk.options.DataflowPipelineOptions;
 import com.google.cloud.dataflow.sdk.options.PipelineOptions;
 import com.google.cloud.dataflow.sdk.options.PipelineOptionsFactory;
 import com.google.cloud.dataflow.sdk.runners.BlockingDataflowPipelineRunner;
-import com.google.cloud.security.scanner.pipelines.OnDemandLiveStateChecker;
-import com.google.cloud.security.scanner.sources.GCSFilesSource;
+import com.google.cloud.security.scanner.pipelines.ExportedServiceAccountKeyRemover;
 import com.google.common.base.Preconditions;
 import java.io.IOException;
-import java.security.GeneralSecurityException;
+import java.io.PrintWriter;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
-/**
- * For running the Dataflow pipeline directly using the Dataflow SDK.
- */
-public class LiveStateCheckerRunner {
+/** Handler class for the Dataflow local runner test endpoint. */
+public class UserManagedKeysApp extends HttpServlet {
 
-  /**
-   * Main function for the runner.
-   * @param args The args this program was called with.
-   * @throws IOException Thrown if there's an error reading from one of the APIs.
-   */
-  public static void main(String[] args) throws IOException, GeneralSecurityException{
-    String org = System.getenv("POLICY_SCANNER_ORG_NAME");
-    String inputRepositoryUrl = System.getenv("POLICY_SCANNER_INPUT_REPOSITORY_URL");
+  @Override
+  public void doGet(HttpServletRequest req, HttpServletResponse resp)
+      throws IOException {
+    PrintWriter out = resp.getWriter();
+    String orgId = System.getenv("POLICY_SCANNER_ORG_ID");
     String sinkUrl = System.getenv("POLICY_SCANNER_SINK_URL");
     String dataflowTmpBucket = System.getenv("POLICY_SCANNER_DATAFLOW_TMP_BUCKET");
     String stagingLocation = "gs://" + dataflowTmpBucket + "/dataflow_tmp";
     boolean executeOnCloud = Boolean.valueOf(System.getenv("POLICY_SCANNER_EXECUTE_ON_CLOUD"));
 
-    Preconditions.checkNotNull(org);
-    Preconditions.checkNotNull(inputRepositoryUrl);
+    Preconditions.checkNotNull(orgId);
     Preconditions.checkNotNull(sinkUrl);
     Preconditions.checkNotNull(dataflowTmpBucket);
-    GCSFilesSource source;
-    try {
-      source = new GCSFilesSource(inputRepositoryUrl, org);
-    } catch (GeneralSecurityException e) {
-      throw new IOException("SecurityException: Cannot create GCSFileSource");
-    }
+
     PipelineOptions options;
     if (executeOnCloud) {
       options = getCloudExecutionOptions(stagingLocation);
     } else {
       options = getLocalExecutionOptions();
     }
-    new OnDemandLiveStateChecker(options, source)
-        .attachSink(TextIO.Write.named("Write messages to GCS").to(sinkUrl))
+
+    new ExportedServiceAccountKeyRemover(options, orgId)
+        .attachSink(TextIO.Write.named("Write output messages").to(sinkUrl))
         .run();
+    out.println("Test passed! The output was written to GCS");
   }
 
-  private static PipelineOptions getCloudExecutionOptions(String stagingLocation) {
+  private PipelineOptions getCloudExecutionOptions(String stagingLocation) {
     DataflowPipelineOptions options = PipelineOptionsFactory.as(DataflowPipelineOptions.class);
     options.setProject(SystemProperty.applicationId.get());
     options.setStagingLocation(stagingLocation);
@@ -75,7 +68,7 @@ public class LiveStateCheckerRunner {
     return options;
   }
 
-  private static PipelineOptions getLocalExecutionOptions() {
+  private PipelineOptions getLocalExecutionOptions() {
     return PipelineOptionsFactory.create();
   }
 }
